backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pycryptodome
